Refactor to modify enc script w/ file_line
This commit changes the implementation of the console env enc script to
modify the shipped ENC script from Puppet Enterprise, piping the output
from the curl command into an awk filter.

Awk was chosen since it is a standard Posix filter application that's
small and fast. Anything that natively supports YAML is probably too
heavyweight to really want to invoke as an ENC if we don't have to.

Modifying the existing script was chosen as a methodology since we don't
necessarily know the details necessary to connect to the console as the
installer has laid down, and this approach lets us not touch that.

An alternate methodology that we COULD use instead would be to control
the external_nodes= configuration line in the puppet.conf file. Not
ready to decide on that vs. adjusting the existing ENC script yet.
Decision for another day.